Added the -vbr option to ignore partition table
Volume Boot Records usually don't contain partition tables, so I've
added the -vbr option to ignore the space reserved for a partition table
and extend the space for the command line. Useful if you don't partition
your drives and just install the system on /dev/sda, or if you have
another bootloader in your MBR (e.g. to select which OS to boot).